requirements.txt
python-dateutil
PySimpleGUI
pyyaml